Refactored code and changed routes
Flattened nested if statements for better readability
Changed GetUser to get all epics (with members) and stories
Changed GetEpics to GetEpic, which gets members and stories
Changed GetStories to GetStory, which gets a single story

+++ models/epic.go
@@ -5,6 +5,7 @@ import (
     "ScrumifyBackend/utils"
     "strconv"
     "time"
+    "database/sql"
 )
 
 type Epic struct {
@@ -12,7 +13,7 @@ type Epic struct {
     Name      string    `db:"name" json:"name"`
     Members   []User    `db:"-" json:"members"`
     Stories   []Story   `db:"-" json:"stories"`
-    CreatedAt time.Time `db:"created_at" json:"created_at"`
+    CreatedAt time.Time `db:"created_at" json:"-"`
     UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
 }
 
@@ -28,12 +29,36 @@ func SetEpicProperties(table *gorp.TableMap) {
         "CURRENT_TIMESTAMP")
 }
 
-func GetEpics(user_id string) ([]Epic, error) {
-    var epics []Epic
-    _, err := Dbmap.Select(&epics, "SELECT * FROM Epic WHERE id IN (SELECT epic_id FROM EpicUserMap WHERE user_id=?)",
-        user_id)
-    utils.PrintErr(err, "GetEpics: Failed to select epics for user " + user_id)
-    return epics, err
+func GetEpic(epic_id string) (Epic, error) {
+    trans, err := Dbmap.Begin()
+    if err != nil {
+        utils.PrintErr(err, "GetEpic: Failed to begin transaction")
+        return Epic{}, err
+    }
+
+    var epic Epic
+    if err := trans.SelectOne(&epic, "SELECT * FROM Epic WHERE id=?", epic_id); err != nil {
+        if err == sql.ErrNoRows {
+            trans.Rollback()
+            return Epic{}, utils.EpicDoesntExist
+        } else {
+            trans.Rollback()
+            utils.PrintErr(err, "GetEpic: Failed to select epic " + epic_id)
+            return Epic{}, err
+        }
+    }
+    if _, err = trans.Select(&epic.Stories, "SELECT * FROM Story WHERE epic_id=?", epic_id); err != nil {
+        trans.Rollback()
+        utils.PrintErr(err, "GetEpic: Failed to select stories for epic " + epic_id)
+        return Epic{}, err
+    }
+    if _, err = trans.Select(&epic.Members, "SELECT * FROM User WHERE id IN (SELECT user_id FROM EpicUserMap WHERE " +
+        "epic_id=?)", epic_id); err != nil {
+        trans.Rollback()
+        utils.PrintErr(err, "GetEpic: Failed to select members for epic " + epic_id)
+        return Epic{}, err
+    }
+    return epic, trans.Commit()
 }
 
 func CreateEpic(user_id string, epic Epic) (Epic, error) {
@@ -43,24 +68,22 @@ func CreateEpic(user_id string, epic Epic) (Epic, error) {
         return Epic{}, err
     }
 
-    if err = trans.Insert(&epic); err == nil {
-        if _, err = trans.Exec("INSERT INTO EpicUserMap (user_id, epic_id) VALUES (?, ?)", user_id, epic.Id);
-            err == nil {
-            var check Epic
-            if err = trans.SelectOne(&check, "SELECT * FROM Epic WHERE id=?", epic.Id); err == nil {
-                return check, trans.Commit()
-            } else {
-                return epic, trans.Commit()
-            }
-        } else {
-            utils.PrintErr(err, "CreateEpic: Failed to insert mapping user_id: " + user_id + " epic_id: " +
-                strconv.FormatInt(epic.Id, 10))
-        }
-    } else {
-        utils.PrintErr(err, "CreateEpic: Failed to insert epic " + strconv.FormatInt(epic.Id, 10))
+    if err = trans.Insert(&epic); err != nil {
+        trans.Rollback()
+        utils.PrintErr(err, "CreateEpic: Failed to insert epic")
+        return Epic{}, err
     }
-    trans.Rollback()
-    return Epic{}, err
+    if _, err = trans.Exec("INSERT INTO EpicUserMap (user_id, epic_id) VALUES (?, ?)", user_id, epic.Id); err != nil {
+        trans.Rollback()
+        utils.PrintErr(err, "CreateEpic: Failed to insert mapping for user_id " + user_id)
+        return Epic{}, err
+    }
+    var check Epic
+    if err = trans.SelectOne(&check, "SELECT * FROM Epic WHERE id=?", epic.Id); err == nil {
+        return check, trans.Commit()
+    } else {
+        return epic, trans.Commit()
+    }
 }
 
 func UpdateEpic(epic Epic) (Epic, error) {
@@ -70,24 +93,23 @@ func UpdateEpic(epic Epic) (Epic, error) {
         return Epic{}, err
     }
 
-    if _, err = trans.Update(&epic); err == nil {
-        var check Epic
-        if err = trans.SelectOne(&check, "SELECT * FROM Epic WHERE id=?", epic.Id); err == nil {
-            return check, trans.Commit()
-        } else {
-            return epic, trans.Commit()
-        }
-    } else {
+    if _, err = trans.Update(&epic); err != nil {
         trans.Rollback()
         utils.PrintErr(err, "UpdateEpic: Failed to update epic " + strconv.FormatInt(epic.Id, 10))
         return Epic{}, err
+    }
+    var check Epic
+    if err = trans.SelectOne(&check, "SELECT * FROM Epic WHERE id=?", epic.Id); err == nil {
+        return check, trans.Commit()
+    } else {
+        return epic, trans.Commit()
     }
 }
 
 func DeleteEpic(mapping EpicUserMap) error {
     _, err := Dbmap.Delete(&mapping)
-    utils.PrintErr(err, "DeleteEpic: Failed to delete mapping user_id: " + strconv.FormatInt(mapping.UserId, 10) +
-        " epic_id: " + strconv.FormatInt(mapping.EpicId, 10))
+    utils.PrintErr(err, "DeleteEpic: Failed to delete mapping for user_id " + strconv.FormatInt(mapping.UserId, 10) +
+        " and epic_id " + strconv.FormatInt(mapping.EpicId, 10))
     if err == nil {
         go removeUnownedEpic(mapping.EpicId)
     }
@@ -106,22 +128,21 @@ func removeUnownedEpic(epic_id int64) {
         return
     }
 
-    var count int64
-    if count, err = trans.SelectInt("SELECT COUNT(*) FROM EpicUserMap WHERE epic_id=?", epic_id);
-        err == nil && count == 0 {
-        epicToBeDeleted := Epic{Id: epic_id}
-        if _, err = trans.Delete(&epicToBeDeleted); err == nil {
-            if _, err = trans.Exec("DELETE FROM Story WHERE epic_id=?", epic_id); err == nil {
-                trans.Commit()
-            } else {
-                utils.PrintErr(err, "removeUnownedEpic: Failed to delete stories for epic " +
-                    strconv.FormatInt(epic_id, 10))
-            }
-        } else {
+    if count, err := trans.SelectInt("SELECT COUNT(*) FROM EpicUserMap WHERE epic_id=?", epic_id); err == nil &&
+        count == 0 {
+        if _, err = trans.Exec("DELETE FROM Epic WHERE id=?", epic_id); err != nil {
+            trans.Rollback()
             utils.PrintErr(err, "removeUnownedEpic: Failed to delete epic " + strconv.FormatInt(epic_id, 10))
         }
+        if _, err = trans.Exec("DELETE FROM Story WHERE epic_id=?", epic_id); err == nil {
+            trans.Commit()
+        } else {
+            trans.Rollback()
+            utils.PrintErr(err, "removeUnownedEpic: Failed to delete stories for epic " +
+                strconv.FormatInt(epic_id, 10))
+        }
     } else {
+        trans.Rollback()
         utils.PrintErr(err, "removeUnownedEpic: Failed to select mappings for epic " + strconv.FormatInt(epic_id, 10))
     }
-    trans.Rollback()
 }
